+++ src/onegov.activity/onegov/activity/__init__.py
@@ -1,6 +1,13 @@
-from onegov.activity.models import Activity, Booking, Occasion, Period
+from onegov.activity.models import (
+    Activity,
+    Attendee,
+    Booking,
+    Occasion,
+    Period
+)
 from onegov.activity.collections import (
     ActivityCollection,
+    AttendeeCollection,
     BookingCollection,
     OccasionCollection,
     PeriodCollection
@@ -9,10 +16,12 @@ from onegov.activity.collections import (
 
 __all__ = [
     'Activity',
+    'Attendee',
     'Booking',
     'Occasion',
     'Period',
     'ActivityCollection',
+    'AttendeeCollection',
     'BookingCollection',
     'OccasionCollection',
     'PeriodCollection'