+++ src/onegov.activity/onegov/activity/collections/__init__.py
@@ -1,10 +1,12 @@
 from onegov.activity.collections.activity import ActivityCollection
+from onegov.activity.collections.attendee import AttendeeCollection
 from onegov.activity.collections.booking import BookingCollection
 from onegov.activity.collections.occasion import OccasionCollection
 from onegov.activity.collections.period import PeriodCollection
 
 __all__ = [
     'ActivityCollection',
+    'AttendeeCollection',
     'BookingCollection',
     'OccasionCollection',
     'PeriodCollection'