+++ src/onegov.activity/onegov/activity/models/__init__.py
@@ -1,4 +1,5 @@
 from onegov.activity.models.activity import Activity, ACTIVITY_STATES
+from onegov.activity.models.attendee import Attendee
 from onegov.activity.models.booking import Booking
 from onegov.activity.models.occasion import Occasion, DAYS
 from onegov.activity.models.period import Period
@@ -6,6 +7,7 @@ from onegov.activity.models.period import Period
 
 __all__ = [
     'Activity',
+    'Attendee',
     'Booking',
     'Occasion',
     'Period',